+++ lib/src/widgets/solid_login.dart
@@ -40,27 +40,20 @@ import 'package:solidpod/solidpod.dart'
         generateDefaultFiles,
         generateCustomFolders,
         setAppDirName;
-import 'package:url_launcher/url_launcher.dart';
 
 import 'package:solidui/src/constants/solid_config.dart';
 import 'package:solidui/src/models/snackbar_config.dart';
-import 'package:solidui/src/widgets/solid_login_auth_handler.dart';
-import 'package:solidui/src/widgets/solid_login_buttons.dart';
+import 'package:solidui/src/widgets/solid_login_build_helper.dart';
 import 'package:solidui/src/widgets/solid_login_helper.dart';
 import 'package:solidui/src/widgets/solid_login_panel.dart';
+import 'package:solidui/src/widgets/solid_login_snackbar_helper.dart';
+import 'package:solidui/src/widgets/solid_login_theme_helper.dart';
 import 'package:solidui/src/widgets/solid_theme_notifier.dart';
 
 /// A widget to login to a Solid server for a user's token to access their POD.
-///
-/// The login screen will be the initial screen of the app when access to the
-/// user's POD is required when the app requires access to the user's POD for
-/// any of its functionality.
 
 class SolidLogin extends StatefulWidget {
-  /// Parameters for authenticating to the Solid server.
-
   const SolidLogin({
-    // Include the literals here so that they are exposed through the docs.
     required this.child,
     this.required = false,
     this.appDirectory = '',
@@ -68,10 +61,8 @@ class SolidLogin extends StatefulWidget {
       'assets/images/default_image.jpg',
       package: 'solidpod',
     ),
-    this.logo = const AssetImage(
-      'assets/images/default_logo.png',
-      package: 'solidpod',
-    ),
+    this.logo =
+        const AssetImage('assets/images/default_logo.png', package: 'solidpod'),
     this.title = 'Log in to your Solid Pod',
     this.webID = SolidConfig.defaultServerUrl,
     this.link = 'https://solidproject.org',
@@ -86,13 +77,10 @@ class SolidLogin extends StatefulWidget {
     super.key,
   });
 
-  /// The app's welcome image used as the left panel or the background.
-  ///
-  /// For a desktop dimensions the image is displayed as the left panel on the
-  /// login screen.  For mobile dimensions (narrow screen) the image forms the
-  /// background behind the Login panel.
+  /// The app's welcome image used as the left panel or the background, and
+  /// the app's logo as displayed at the top of the login panel.
 
-  final AssetImage image;
+  final AssetImage image, logo;
 
   /// The style of the REGISTER button.
 
@@ -114,23 +102,12 @@ class SolidLogin extends StatefulWidget {
 
   final ChangeKeyButtonStyle changeKeyButtonStyle;
 
-  /// The app's logo as displayed at the top of the login panel.
+  /// The login text indicating what we are loging in to, the URI of the
+  /// user's webID used to identify the Solid server to authenticate against,
+  /// and the URL used as the value of the Visit link. Visit the link by
+  /// clicking info button.
 
-  final AssetImage logo;
-
-  /// The login text indicating what we are loging in to.
-
-  final String title;
-
-  /// The URI of the user's webID used to identify the Solid server to
-  /// authenticate against.
-
-  final String webID;
-
-  /// The URL used as the value of the Visit link. Visit the link by clicking
-  /// info button.
-
-  final String link;
+  final String title, webID, link;
 
   /// The child widget after logging in.
 
@@ -157,14 +134,7 @@ class SolidLogin extends StatefulWidget {
   final SnackbarConfig snackbarConfig;
 
   /// Custom list of folders to be created inside the data folder.
-  /// Following are few examples.
-  ///   - Custom folder 'myDir1' will be created as '/data/myDir1'
-  ///   - Custom folder 'data' will be created as '/data/data'
-  /// If multi-level folder structure is needed you need to provide
-  /// upper level folders first in the list. For instance, to create
-  /// 'myDir1/myDir2/myDir3', add three values to the list as follows
-  /// in that order.
-  /// 'myDir1', 'myDir1/myDir2', 'myDir1/myDir2/myDir3'
+
   final List customFolderPathList;
 
   @override
@@ -172,14 +142,11 @@ class SolidLogin extends StatefulWidget {
 }
 
 class _SolidLoginState extends State<SolidLogin> with WidgetsBindingObserver {
-  // This strings will hold the application version number and app name.
-  // Initially, it's an empty string because the actual version number
-  // will be obtained asynchronously from the app's package information.
+  /// The app version and the app name.
 
-  String appVersion = '';
-  String appName = '';
+  String appVersion = '', appName = '';
 
-  // Check whether the dialog was dismissed by the user.
+  /// Check whether the dialog was dismissed by the user.
 
   bool isDialogCanceled = false;
 
@@ -192,13 +159,11 @@ class _SolidLoginState extends State<SolidLogin> with WidgetsBindingObserver {
   Map<dynamic, dynamic> defaultFiles = {};
 
   // Focus nodes for keyboard navigation.
-  // Tab order: login -> continue -> register -> info -> server input.
 
-  late final FocusNode _loginFocusNode;
-  late final FocusNode _continueFocusNode;
-  late final FocusNode _registerFocusNode;
-  late final FocusNode _infoFocusNode;
-  late final FocusNode _serverInputFocusNode;
+  late final FocusNode _loginFocusNode, _continueFocusNode;
+  late final FocusNode _registerFocusNode,
+      _infoFocusNode,
+      _serverInputFocusNode;
 
   @override
   void initState() {
@@ -247,62 +212,22 @@ class _SolidLoginState extends State<SolidLogin> with WidgetsBindingObserver {
 
   /// Callback when theme notifier changes.
 
-  void _onThemeChanged() {
-    if (mounted) {
-      setState(() {});
-    }
-  }
-
-  /// Determines if dark mode should be used based on the current theme mode.
-  /// When in system mode, follows the system brightness.
-  /// When explicitly set to light or dark, uses that mode.
-
-  bool get isDarkMode {
-    switch (solidThemeNotifier.themeMode) {
-      case ThemeMode.system:
-        return MediaQuery.platformBrightnessOf(context) == Brightness.dark;
-      case ThemeMode.light:
-        return false;
-      case ThemeMode.dark:
-        return true;
-    }
-  }
-
-  // Fetch the package information.
+  void _onThemeChanged() => mounted ? setState(() {}) : null;
+  bool get isDarkMode => SolidLoginThemeHelper.isDarkMode(context);
 
   Future<void> _initPackageInfo() async {
-    // Check if widget is still mounted before starting any async operations.
-    // This prevents unnecessary work if the widget has been disposed.
-
     if (!mounted) return;
-
     await setAppDirName(widget.appDirectory);
     final folders = await generateDefaultFolders();
     final files = await generateDefaultFiles();
-
     final customFolders = generateCustomFolders(widget.customFolderPathList);
-
-    // Check if widget is still mounted after async operations and before setState.
-    // This prevents "setState() called after dispose()" errors that can occur
-    // if the widget was disposed while async operations were running.
-
     if (!mounted) return;
-
     setState(() {
       defaultFolders = folders + customFolders;
       defaultFiles = files;
     });
-
-    // Fetch the app information.
-
     final appInfo = await getAppNameVersion();
-
-    // Check if widget is still mounted after final async operation and before setState.
-    // This ensures we don't call setState on a disposed widget, which would throw
-    // a FlutterError and potentially crash the app.
-
     if (!mounted) return;
-
     setState(() {
       appName = appInfo.name;
       appVersion = appInfo.version;
@@ -312,11 +237,7 @@ class _SolidLoginState extends State<SolidLogin> with WidgetsBindingObserver {
   // Function to update [_isDialogCanceled].
 
   void updateState() {
-    if (mounted) {
-      setState(() {
-        isDialogCanceled = true;
-      });
-    }
+    if (mounted) setState(() => isDialogCanceled = true);
   }
 
   // Helper method to create and show a snackbar with consistent theming.
@@ -329,135 +250,62 @@ class _SolidLoginState extends State<SolidLogin> with WidgetsBindingObserver {
     final currentTheme = isDarkMode
         ? widget.themeConfig.darkTheme
         : widget.themeConfig.lightTheme;
-
-    final backgroundColor = widget.snackbarConfig.backgroundColor ??
-        (isDarkMode
-            ? currentTheme.backgroundColor.withValues(alpha: 0.9)
-            : currentTheme.backgroundColor.withValues(alpha: 0.7));
-
-    ScaffoldMessenger.of(context).showSnackBar(
-      SnackBar(
-        content: Text(
-          message,
-          style: TextStyle(
-            color: widget.snackbarConfig.textColor != Colors.black
-                ? widget.snackbarConfig.textColor
-                : currentTheme.textColor,
-            fontWeight: FontWeight.w500,
-          ),
-        ),
-        duration: duration ?? widget.snackbarConfig.duration,
-        behavior: SnackBarBehavior.floating,
-        backgroundColor: backgroundColor,
-        shape: RoundedRectangleBorder(
-          borderRadius: BorderRadius.circular(
-            widget.snackbarConfig.borderRadius,
-          ),
-          side: BorderSide(color: currentTheme.dividerColor, width: 0.5),
-        ),
-        action: showAction
-            ? SnackBarAction(
-                label: 'OK',
-                textColor: widget.snackbarConfig.actionTextColor != Colors.black
-                    ? widget.snackbarConfig.actionTextColor
-                    : currentTheme.titleColor,
-                onPressed: () {},
-              )
-            : null,
-      ),
+    SolidLoginSnackbarHelper.showSnackbar(
+      context,
+      message: message,
+      isDarkMode: isDarkMode,
+      currentTheme: currentTheme,
+      snackbarConfig: widget.snackbarConfig,
+      duration: duration,
+      showAction: showAction,
     );
   }
 
-  // Toggle between light and dark mode using the global theme notifier.
-  // This ensures consistency with the rest of the app.
-
-  void _toggleTheme() {
-    solidThemeNotifier.toggleTheme();
-  }
+  void _toggleTheme() => solidThemeNotifier.toggleTheme();
 
   @override
   Widget build(BuildContext context) {
-    // Use the internal state for theme instead of system brightness.
-
     final currentTheme = isDarkMode
         ? widget.themeConfig.darkTheme
         : widget.themeConfig.lightTheme;
 
-    // The login box's default image Widget for the left/background panel
-    // depending on screen width.
-
     final loginBoxDecor = BoxDecoration(
       image: DecorationImage(image: widget.image, fit: BoxFit.cover),
     );
-
-    // Text controller for the URI of the solid server to which an authenticate
-    // request is sent.
-
     final webIdController = TextEditingController()..text = widget.webID;
 
-    // Build all buttons using the button builder.
-    // User input from text field will override the default server URL.
-
-    final registerButton = FocusTraversalOrder(
-      order: const NumericFocusOrder(3),
-      child: SolidLoginButtons.buildRegisterButton(
-        style: widget.registerButtonStyle,
-        onPressed: () {
-          final webId = webIdController.text.trim().isNotEmpty
-              ? webIdController.text.trim()
-              : SolidConfig.defaultServerUrl;
-          launchUrl(Uri.parse('$webId/.account/login/password/register/'));
-        },
-        focusNode: _registerFocusNode,
-      ),
+    final registerButton = SolidLoginBuildHelper.buildRegisterButton(
+      style: widget.registerButtonStyle,
+      webIdController: webIdController,
+      focusNode: _registerFocusNode,
     );
 
-    final loginButton = FocusTraversalOrder(
-      order: const NumericFocusOrder(1),
-      child: SolidLoginButtons.buildLoginButton(
-        style: widget.loginButtonStyle,
-        onPressed: () async {
-          final podServer = webIdController.text.trim().isNotEmpty
-              ? webIdController.text.trim()
-              : SolidConfig.defaultServerUrl;
-
-          isDialogCanceled = false;
-          await SolidLoginAuthHandler.handleLogin(
-            context: context,
-            podServer: podServer,
-            defaultFolders: defaultFolders,
-            defaultFiles: defaultFiles,
-            originalLoginWidget: widget,
-            childWidget: widget.child,
-            isDialogCanceled: isDialogCanceled,
-            updateDialogCanceledState: updateState,
-            showSnackbar: _showSnackbar,
-          );
-        },
-        focusNode: _loginFocusNode,
-        autofocus: true,
-      ),
+    final loginButton = SolidLoginBuildHelper.buildLoginButton(
+      context: context,
+      style: widget.loginButtonStyle,
+      webIdController: webIdController,
+      defaultFolders: defaultFolders,
+      defaultFiles: defaultFiles,
+      originalWidget: widget,
+      childWidget: widget.child,
+      getIsDialogCanceled: () => isDialogCanceled,
+      updateDialogCanceledState: updateState,
+      showSnackbar: _showSnackbar,
+      focusNode: _loginFocusNode,
     );
 
-    final continueButton = FocusTraversalOrder(
-      order: const NumericFocusOrder(2),
-      child: SolidLoginButtons.buildContinueButton(
-        style: widget.continueButtonStyle,
-        onPressed: () async => await pushReplacement(context, widget.child),
-        focusNode: _continueFocusNode,
-      ),
+    final continueButton = SolidLoginBuildHelper.buildContinueButton(
+      context: context,
+      style: widget.continueButtonStyle,
+      childWidget: widget.child,
+      focusNode: _continueFocusNode,
     );
 
-    final infoButton = FocusTraversalOrder(
-      order: const NumericFocusOrder(4),
-      child: SolidLoginButtons.buildInfoButton(
-        style: widget.infoButtonStyle,
-        link: widget.link,
-        focusNode: _infoFocusNode,
-      ),
+    final infoButton = SolidLoginBuildHelper.buildInfoButton(
+      style: widget.infoButtonStyle,
+      link: widget.link,
+      focusNode: _infoFocusNode,
     );
-
-    // Build the login panel content.
 
     final loginPanelContent = SolidLoginPanel.buildPanelContent(
       context: context,
@@ -474,15 +322,11 @@ class _SolidLoginState extends State<SolidLogin> with WidgetsBindingObserver {
       serverInputFocusNode: _serverInputFocusNode,
     );
 
-    // Add theme toggle to the panel.
-
     final loginPanelDecor = SolidLoginPanel.buildPanelWithThemeToggle(
       panelContent: loginPanelContent,
       currentThemeMode: solidThemeNotifier.themeMode,
       onThemeToggle: _toggleTheme,
     );
-
-    // Build the complete login panel.
 
     final loginPanel = SolidLoginPanel.buildCompletePanel(
       context: context,
@@ -490,37 +334,10 @@ class _SolidLoginState extends State<SolidLogin> with WidgetsBindingObserver {
       currentTheme: currentTheme,
     );
 
-    // Build and return the final Scaffold.
-    // Wrap with FocusTraversalGroup to enable ordered keyboard navigation.
-
-    return Scaffold(
-      body: FocusTraversalGroup(
-        policy: OrderedTraversalPolicy(),
-        child: GestureDetector(
-          onTap: () {
-            ScaffoldMessenger.of(context).hideCurrentSnackBar();
-          },
-          behavior: HitTestBehavior.deferToChild,
-          child: SafeArea(
-            child: DecoratedBox(
-              decoration: isNarrowScreen(context)
-                  ? loginBoxDecor
-                  : const BoxDecoration(),
-              child: Row(
-                children: [
-                  isNarrowScreen(context)
-                      ? Container()
-                      : Expanded(
-                          flex: 7,
-                          child: Container(decoration: loginBoxDecor),
-                        ),
-                  Expanded(flex: 5, child: loginPanel),
-                ],
-              ),
-            ),
-          ),
-        ),
-      ),
+    return SolidLoginBuildHelper.buildScaffold(
+      context: context,
+      loginBoxDecor: loginBoxDecor,
+      loginPanel: loginPanel,
     );
   }
 }
