+++ lib/src/widgets/solid_nav_drawer.dart
@@ -32,18 +32,13 @@ library;
 
 import 'package:flutter/material.dart';
 
-import 'package:gap/gap.dart';
 import 'package:package_info_plus/package_info_plus.dart';
-import 'package:url_launcher/url_launcher.dart';
-import 'package:version_widget/version_widget.dart';
 
 import 'package:solidui/src/constants/navigation.dart';
+import 'package:solidui/src/widgets/solid_nav_drawer_header.dart';
 import 'package:solidui/src/widgets/solid_nav_models.dart';
 
 /// A solid navigation drawer component.
-///
-/// This widget provides a collapsible navigation drawer that displays
-/// when the screen is narrow, replacing the navigation rail.
 
 class SolidNavDrawer extends StatefulWidget {
   /// User information to display in the drawer header.
@@ -143,6 +138,8 @@ class _SolidNavDrawerState extends State<SolidNavDrawer> {
     return '0.0.0+0';
   }
 
+  bool _canLogout() => widget.showLogout && widget.onLogout != null;
+
   @override
   Widget build(BuildContext context) {
     final theme = Theme.of(context);
@@ -158,81 +155,28 @@ class _SolidNavDrawerState extends State<SolidNavDrawer> {
       child: ListView(
         padding: EdgeInsets.only(top: MediaQuery.of(context).padding.top),
         children: <Widget>[
-          // User info header (if provided).
-          if (widget.userInfo != null) _buildUserInfoHeader(context, theme),
-
-          // Navigation items.
+          if (widget.userInfo != null)
+            SolidNavDrawerHeader.build(
+              context: context,
+              theme: theme,
+              user: widget.userInfo!,
+              isVersionLoaded: _isVersionLoaded,
+              appVersion: _appVersion,
+              getVersionToDisplay: _getVersionToDisplay,
+            ),
           Container(
             padding: const EdgeInsets.all(NavigationConstants.navDrawerPadding),
             child: Column(
               children: [
-                // Main navigation tabs.
                 ...widget.tabs.asMap().entries.map((entry) {
                   final index = entry.key;
                   final tab = entry.value;
-
-                  return ListTile(
-                    leading: Icon(
-                      tab.icon,
-                      color: index == widget.selectedIndex
-                          ? theme.colorScheme.primary
-                          : theme.colorScheme.onSurface.withValues(alpha: 0.7),
-                    ),
-                    title: Text(
-                      tab.title,
-                      style: TextStyle(
-                        fontWeight: index == widget.selectedIndex
-                            ? FontWeight.w600
-                            : FontWeight.w400,
-                        color: index == widget.selectedIndex
-                            ? theme.colorScheme.primary
-                            : theme.colorScheme.onSurface,
-                      ),
-                    ),
-                    selected: index == widget.selectedIndex,
-                    selectedTileColor: theme.colorScheme.primary.withValues(
-                      alpha: 0.1,
-                    ),
-                    onTap: () {
-                      widget.onTabSelected(index);
-                      Navigator.of(context).pop(); // Close drawer.
-                    },
-                  );
+                  return _buildNavTile(context, theme, index, tab);
                 }),
-
-                // Additional menu items (if provided).
                 if (widget.additionalMenuItems != null)
                   ...widget.additionalMenuItems!,
-
-                // Divider and logout option.
-                if (widget.showLogout && widget.onLogout != null) ...[
-                  Divider(
-                    height: NavigationConstants.navDividerHeight,
-                    color: theme.dividerColor,
-                  ),
-                  ListTile(
-                    leading: Icon(
-                      widget.logoutIcon ?? Icons.logout,
-                      color: _canLogout()
-                          ? theme.colorScheme.error
-                          : theme.disabledColor,
-                    ),
-                    title: Text(
-                      widget.logoutText ?? 'Logout',
-                      style: TextStyle(
-                        color: _canLogout()
-                            ? theme.colorScheme.error
-                            : theme.disabledColor,
-                      ),
-                    ),
-                    onTap: _canLogout()
-                        ? () {
-                            Navigator.of(context).pop(); // Close drawer first.
-                            widget.onLogout!(context);
-                          }
-                        : null,
-                  ),
-                ],
+                if (widget.showLogout && widget.onLogout != null)
+                  ..._buildLogoutSection(context, theme),
               ],
             ),
           ),
@@ -241,219 +185,62 @@ class _SolidNavDrawerState extends State<SolidNavDrawer> {
     );
   }
 
-  Widget _buildUserInfoHeader(BuildContext context, ThemeData theme) {
-    final user = widget.userInfo!;
-    final bool willShowVersion = user.versionConfig != null;
-    final double bottomPadding = willShowVersion
-        ? 8.0 // Add spacing below version
-        : NavigationConstants.userHeaderBottomPadding;
-
-    return Container(
-      padding: EdgeInsets.only(
-        top: NavigationConstants.userHeaderTopPadding +
-            MediaQuery.of(context).padding.top,
-        bottom: bottomPadding,
-      ),
-      decoration: BoxDecoration(color: theme.colorScheme.primaryContainer),
-      child: Column(
-        children: [
-          // User avatar.
-          user.avatar ??
-              Icon(
-                user.avatarIcon ?? Icons.account_circle,
-                size: user.avatarSize ?? NavigationConstants.userAvatarSize,
-                color: theme.colorScheme.onPrimaryContainer,
-              ),
-
-          const Gap(NavigationConstants.userInfoSpacing),
-
-          // User name.
-          Text(
-            user.effectiveUserName,
-            style: TextStyle(
-              color: theme.colorScheme.onPrimaryContainer,
-              fontSize: NavigationConstants.userNameFontSize,
-              fontWeight: FontWeight.w600,
-            ),
-          ),
-
-          // WebID (if enabled and available).
-          if (user.showWebId &&
-              user.webId != null &&
-              user.webId!.isNotEmpty) ...[
-            const Gap(NavigationConstants.webIdSpacing),
-            Container(
-              padding: const EdgeInsets.symmetric(
-                horizontal: NavigationConstants.webIdHorizontalPadding,
-              ),
-              child: InkWell(
-                onTap: () => _launchProfileUrl(user.webId!),
-                child: Text(
-                  _getSimplifiedUrl(user.webId!),
-                  style: TextStyle(
-                    color: theme.colorScheme.onPrimaryContainer.withValues(
-                      alpha: 0.8,
-                    ),
-                    fontSize: NavigationConstants.webIdFontSize,
-                  ),
-                  textAlign: TextAlign.center,
-                ),
-              ),
-            ),
-          ],
-
-          if (user.versionConfig != null) ...[
-            const Gap(NavigationConstants.webIdSpacing),
-            if (_isVersionLoaded &&
-                _appVersion != null &&
-                _appVersion!.isNotEmpty)
-              _buildVersionInfo(context, theme, user.versionConfig!)
-            else
-              // The offset of drawer menu entry background block.
-
-              const SizedBox(height: 23.0),
-          ],
-        ],
-      ),
-    );
-  }
-
-  /// Builds the version information widget.
-
-  Widget _buildVersionInfo(
+  Widget _buildNavTile(
     BuildContext context,
     ThemeData theme,
-    SolidVersionConfig versionConfig,
+    int index,
+    SolidNavTab tab,
   ) {
-    final versionString =
-        (versionConfig.version != null && versionConfig.version!.isNotEmpty)
-            ? versionConfig.version!
-            : _getVersionToDisplay();
-
-    return VersionWidget(
-      version: versionString,
-      changelogUrl: versionConfig.changelogUrl,
-      showDate: versionConfig.showDate,
-      userTextStyle: versionConfig.userTextStyle,
+    return ListTile(
+      leading: Icon(
+        tab.icon,
+        color: index == widget.selectedIndex
+            ? theme.colorScheme.primary
+            : theme.colorScheme.onSurface.withValues(alpha: 0.7),
+      ),
+      title: Text(
+        tab.title,
+        style: TextStyle(
+          fontWeight:
+              index == widget.selectedIndex ? FontWeight.w600 : FontWeight.w400,
+          color: index == widget.selectedIndex
+              ? theme.colorScheme.primary
+              : theme.colorScheme.onSurface,
+        ),
+      ),
+      selected: index == widget.selectedIndex,
+      selectedTileColor: theme.colorScheme.primary.withValues(alpha: 0.1),
+      onTap: () {
+        widget.onTabSelected(index);
+        Navigator.of(context).pop();
+      },
     );
   }
 
-  /// Determines if logout functionality is available.
-
-  bool _canLogout() {
-    // Logout is available if onLogout callback is provided and showLogout is
-    // true.
-
-    return widget.showLogout && widget.onLogout != null;
-  }
-
-  /// Simplifies the WebID URL for display purposes.
-  /// Returns the domain name and username for display.
-
-  String _getSimplifiedUrl(String webId) {
-    try {
-      final uri = Uri.parse(webId);
-
-      // Get the host (server domain).
-
-      String host = uri.host;
-
-      // Extract username from the path.
-
-      String username = '';
-      final pathSegments = uri.pathSegments;
-
-      // Typical webID format: /username/profile/card#me
-      // So the username is usually the first path segment.
-
-      if (pathSegments.isNotEmpty) {
-        username = pathSegments.first;
-      }
-
-      // Return formatted display string.
-
-      if (username.isNotEmpty) {
-        return '$host/$username';
-      } else {
-        // Fallback to just the host if no username found.
-
-        return host;
-      }
-    } catch (e) {
-      // Fallback parsing for malformed URLs.
-
-      try {
-        // Remove common prefixes and suffixes.
-
-        String cleaned = webId;
-
-        // Remove protocol.
-
-        if (cleaned.startsWith('https://')) {
-          cleaned = cleaned.substring(8);
-        } else if (cleaned.startsWith('http://')) {
-          cleaned = cleaned.substring(7);
-        }
-
-        // Remove common webID suffix.
-
-        const suffix = '/profile/card#me';
-        if (cleaned.endsWith(suffix)) {
-          cleaned = cleaned.substring(0, cleaned.length - suffix.length);
-        }
-
-        return cleaned;
-      } catch (e2) {
-        // Final fallback: return original webID.
-
-        return webId;
-      }
-    }
-  }
-
-  /// Gets the complete profile card URL from a WebID.
-
-  String _getProfileCardUrl(String webId) {
-    try {
-      final uri = Uri.parse(webId);
-
-      // Get the scheme, host, and path segments.
-
-      final scheme = uri.scheme;
-      final host = uri.host;
-      final pathSegments = uri.pathSegments;
-
-      // Typical webID format: /username/profile/card#me
-      // We want to construct: https: //host/username/profile/card#
-
-      if (pathSegments.isNotEmpty) {
-        final username = pathSegments.first;
-        return '$scheme:' '//$host/$username/profile/card#';
-      } else {
-        // Fallback: return the original webId.
-
-        return webId;
-      }
-    } catch (e) {
-      // Fallback: return original webID.
-
-      return webId;
-    }
-  }
-
-  /// Launches the profile card URL in a browser.
-
-  Future<void> _launchProfileUrl(String webId) async {
-    try {
-      final profileUrl = _getProfileCardUrl(webId);
-      final uri = Uri.parse(profileUrl);
-      if (await canLaunchUrl(uri)) {
-        await launchUrl(uri, mode: LaunchMode.externalApplication);
-      } else {
-        debugPrint('Cannot launch URL: $profileUrl');
-      }
-    } catch (e) {
-      debugPrint('Error launching profile URL: $e');
-    }
+  List<Widget> _buildLogoutSection(BuildContext context, ThemeData theme) {
+    return [
+      Divider(
+        height: NavigationConstants.navDividerHeight,
+        color: theme.dividerColor,
+      ),
+      ListTile(
+        leading: Icon(
+          widget.logoutIcon ?? Icons.logout,
+          color: _canLogout() ? theme.colorScheme.error : theme.disabledColor,
+        ),
+        title: Text(
+          widget.logoutText ?? 'Logout',
+          style: TextStyle(
+            color: _canLogout() ? theme.colorScheme.error : theme.disabledColor,
+          ),
+        ),
+        onTap: _canLogout()
+            ? () {
+                Navigator.of(context).pop();
+                widget.onLogout!(context);
+              }
+            : null,
+      ),
+    ];
   }
 }
